requirements.txt
Django==5.1
pandas
numpy
pyspark
scikit-learn
joblib
django-ninja
fastapi
uvicorn
